rename handoffService to SharingService
