+++ lib/ag-solo/vats/bootstrap.js
@@ -59,7 +59,7 @@ export default function setup(syscall, state, helpers) {
       function makeBundler(contractHost, vats) {
         return harden({
           async createDemoBundle(nickname) {
-            const handoff = await E(vats.handoff).getSharedHandoffService();
+            const sharingService = await E(vats.sharing).getSharingService();
             const registrar = await E(vats.registrar).getSharedRegistrar();
             const chainTimerService = await E(vats.timer).createTimerService();
             const pixelBundle = await E(vats.pixel).createPixelBundle(nickname);
@@ -68,7 +68,7 @@ export default function setup(syscall, state, helpers) {
             return harden({
               ...pixelBundle,
               chainTimerService,
-              handoff,
+              sharingService,
               moolah,
               contractHost,
               registrar,
