+++ lib/ag-solo/vats/vat-pixel.js
@@ -1,5 +1,5 @@
 import harden from '@agoric/harden';
-import { makeHandoffService } from '@agoric/ertp/more/handoff/handoff';
+import { makeSharingService } from '@agoric/ertp/more/sharing/sharing';
 import { makeGallery } from '@agoric/ertp/more/pixels/gallery';
 import pubsub from './pubsub';
 
@@ -15,11 +15,11 @@ import pubsub from './pubsub';
 // user's nickname, and should create a record of demo objects for them to
 // use (named the "chainBundle"). The client will fetch this record when it
 // starts up, making its properties available to the REPL on the `home`
-// object (e.g. `home.gallery`, `home.handoffService`, etc).
+// object (e.g. `home.gallery`, `home.sharingService`, etc).
 
 function build(E, log) {
   let sharedGalleryUserFacet;
-  let sharedHandoffService;
+  let sharingService;
   let sharedDustAssay;
   let contractHost;
 
@@ -42,17 +42,16 @@ function build(E, log) {
     stateChangeHandler(await E(gallery.readFacet).getState());
     const assays = await E(gallery.userFacet).getAssays();
     sharedDustAssay = assays.dustAssay;
-    sharedHandoffService = makeHandoffService();
+    sharingService = makeSharingService();
     contractHost = host;
   }
 
   async function createPixelBundle(_nickname) {
     const gallery = sharedGalleryUserFacet;
-    const handoffService = sharedHandoffService;
     const purse = await sharedDustAssay.makeEmptyPurse();
     const chainBundle = {
       gallery,
-      handoffService,
+      sharingService,
       purse,
       canvasStatePublisher,
       contractHost,
